--- client/src/components/AppBar/index.js
@@ -1,3 +1,0 @@
-import { AppBar } from './AppBar';
-
-export default AppBar;
