--- client/src/components/Feedback/index.js
@@ -1,3 +1,0 @@
-import { Feedback } from './Feedback';
-
-export default Feedback;
