--- client/src/components/Home/index.js
@@ -1,3 +1,0 @@
-import { Home } from './Home';
-
-export default Home;
