--- client/src/components/WorldWeather/index.js
@@ -1,3 +1,0 @@
-import { WorldWeather } from './WorldWeather';
-
-export default WorldWeather;
