+++ client/src/index.html
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
-<html lang='en'>
-<head>
-  <meta charset='utf-8' />
-  <meta content='width=device-width, initial-scale=1, shrink-to-fit=no' name='viewport' />
-  <title>React Training App</title>
-</head>
-<body>
-<div id='app'></div>
-</body>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta content="width=device-width, initial-scale=1, shrink-to-fit=no" name="viewport" />
+    <title>React Training App</title>
+  </head>
+  <body>
+    <div id="app"></div>
+  </body>
 </html>
