+++ client/src/modules/AboutUs/index.js
@@ -1,0 +1,3 @@
+import { AboutUs } from './components/AboutUs';
+
+export default AboutUs;
