+++ client/src/modules/AppBar/index.js
@@ -1,0 +1,3 @@
+import { AppBar } from './containers/AppBar';
+
+export default AppBar;
