+++ client/src/modules/Feedback/index.js
@@ -1,0 +1,3 @@
+import { Feedback } from './components/Feedback';
+
+export default Feedback;
