+++ client/src/modules/Footer/index.js
@@ -1,0 +1,3 @@
+import { Footer } from './components/Footer';
+
+export default Footer;
