+++ client/src/modules/Home/index.js
@@ -1,0 +1,3 @@
+import { Home } from './components/Home';
+
+export default Home;
