+++ client/src/modules/WorldWeather/index.js
@@ -1,0 +1,3 @@
+import { WorldWeather } from './components/WorldWeather';
+
+export default WorldWeather;
